# zhglxt-file-manager/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: zhglxt File Manager API
  description: |
    Comprehensive file management API with elFinder integration, supporting:
    - Local and cloud storage backends (AWS S3, Alibaba OSS, Tencent COS)
    - File upload, download, delete, move, and preview operations
    - Thumbnail generation and watermarking
    - Advanced security features and permission control
    - Multi-language support (English/Chinese)
  version: 1.0.0
  contact:
    name: zhglxt Development Team
    email: support@zhglxt.com
    url: https://github.com/zhglxt/zhglxt
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: Default server
security:
  - shiroAuth: []
paths:
...
  /file-manager/connector:
    get:
      tags:
        - elFinder Connector
      summary: elFinder connector GET requests
      description: Handle elFinder connector GET requests
      operationId: elFinderConnectorGet
      parameters:
        - name: cmd
          in: query
          required: true
          schema:
            type: string
          description: elFinder command
        - name: target
          in: query
          required: false
          schema:
            type: string
          description: Target hash
      responses:
        '200':
          description: Command executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElFinderResponse'

    post:
      tags:
        - elFinder Connector
      summary: elFinder connector POST requests
      description: Handle elFinder connector POST requests (file uploads, etc.)
      operationId: elFinderConnectorPost
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cmd:
                  type: string
                  description: elFinder command
                target:
                  type: string
                  description: Target hash
                upload:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Files to upload
      responses:
        '200':
          description: Command executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElFinderResponse'
...
components:
  securitySchemes:
    shiroAuth:
      type: http
      scheme: bearer
      description: Shiro-based authentication token
  schemas:
...
    ElFinderResponse:
      type: object
      properties:
        files:
          type: array
          items:
            type: object
        cwd:
          type: object
        options:
          type: object
        error:
          type: array
          items:
            type: string
...